feat: export excel with oauth.

+++ myopia-business/management/src/main/java/com/wupol/myopia/business/management/domain/model/District.java
@@ -40,12 +40,12 @@ public class District implements Serializable {
     /**
      * 行政区代码
      */
-    private Integer code;
+    private Long code;
 
     /**
      * 上级行政区代码（省级统一为100000000000）
      */
-    private Integer parentCode;
+    private Long parentCode;
 
     /**
      * 片区代码
